test: consolidate root package tests by theme into three files
Group the seven root package-main test files into three by concern:
main_test.go (main wiring + dogfood specs-load), schema_test.go (input +
output JSON schema), and drift_test.go (docs, site, and examples
generated-artifact drift guards). No test is added, removed, or changed;
the custom hasPrefix in the former site_test.go is dropped for
strings.HasPrefix now that the merged file already imports strings.
